Split core quality test boundaries
